refactor!: remove transactions and blocks methods from blockfrost wallet provider
add `blockfrostChainHistoryProvider` with transactions and blocks methods

add `chainHistoryProvider` to wallet dependencies
