+++ packages/blockfrost/src/index.ts
@@ -4,5 +4,6 @@ export * from './blockfrostAssetProvider';
 export * from './blockfrostTxSubmitProvider';
 export * from './blockfrostNetworkInfoProvider';
 export * from './blockfrostUtxoProvider';
+export * from './blockfrostChainHistoryProvider';
 export { Options } from '@blockfrost/blockfrost-js/lib/types';
 export { BlockFrostAPI } from '@blockfrost/blockfrost-js';
