+++ packages/blockfrost/test/e2e/config.test.ts
@@ -1,7 +1,9 @@
-import { walletProvider } from './config';
+import { assetProvider, chainHistoryProvider, walletProvider } from './config';
 
 describe('config', () => {
   test('all config variables are set', () => {
     expect(walletProvider).toBeTruthy();
+    expect(assetProvider).toBeTruthy();
+    expect(chainHistoryProvider).toBeTruthy();
   });
 });
