+++ packages/blockfrost/test/e2e/config.ts
@@ -1,5 +1,5 @@
 import { BlockFrostAPI } from '@blockfrost/blockfrost-js';
-import { blockfrostAssetProvider, blockfrostWalletProvider } from '../../src';
+import { blockfrostAssetProvider, blockfrostChainHistoryProvider, blockfrostWalletProvider } from '../../src';
 
 const testnetprojectId = process.env.BLOCKFROST_API_KEY;
 if (!testnetprojectId) throw new Error('BLOCKFROST_API_KEY not set');
@@ -7,3 +7,4 @@ if (!testnetprojectId) throw new Error('BLOCKFROST_API_KEY not set');
 const blockfrostApi = new BlockFrostAPI({ isTestnet: true, projectId: testnetprojectId });
 export const walletProvider = blockfrostWalletProvider(blockfrostApi);
 export const assetProvider = blockfrostAssetProvider(blockfrostApi);
+export const chainHistoryProvider = blockfrostChainHistoryProvider(blockfrostApi);
