+++ packages/blockfrost/test/e2e/queryTransactions.test.ts
@@ -1,10 +1,10 @@
 import { Cardano, InvalidStringError } from '@cardano-sdk/core';
-import { walletProvider } from './config';
+import { chainHistoryProvider } from './config';
 
-describe('blockfrostWalletProvider', () => {
+describe('blockfrostChainHistoryProvider', () => {
   describe('transactionsByHashes', () => {
     it('parses metadata correctly', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('84801fb64a9c5078c406ead24017ba0b069ef6ac6446fef8bdb8f97bade3cfa5')
       ]);
       expect(tx.auxiliaryData!.body.blob!.get(9_223_372_036_854_775_707n)).toEqual(
@@ -12,7 +12,7 @@ describe('blockfrostWalletProvider', () => {
       );
     });
     it('parses collaterals correctly', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('12fa9af65e21b36ec4dc4cbce478e911d52585adb46f2b4fe3d6563e7ee5a61a')
       ]);
       expect(tx.body.inputs!.length).toEqual(1);
@@ -29,7 +29,7 @@ describe('blockfrostWalletProvider', () => {
     // TODO: waiting for release/deploy in db-sync/blockfrost
     // https://github.com/input-output-hk/cardano-db-sync/issues/1019
     it.skip('has collaterals for failed contract', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('9298f499a4c4aeba53a984cb4df0f9a93b7d158da4c2c2d12a06530841f94cd7')
       ]);
       expect(tx.body.inputs!.length).toEqual(0);
@@ -37,7 +37,7 @@ describe('blockfrostWalletProvider', () => {
       expect(tx.body.collaterals!.length).toEqual(1);
     });
     it('has withdrawals', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('e92e3c2ce94d61449876e23ba9170ca20868ee447f13703c5fa7e888cc1701e1')
       ]);
       expect(tx.body.withdrawals).toEqual([
@@ -45,7 +45,7 @@ describe('blockfrostWalletProvider', () => {
       ]);
     });
     it('has redeemer', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('3353309c6d3f2200c4e2084f76c1f1495f00eb21ea62f29c01da2adac71e1068')
       ]);
       expect(tx.witness.redeemers).toEqual([
@@ -59,13 +59,13 @@ describe('blockfrostWalletProvider', () => {
     });
     // TODO: not implemented
     it.skip('has mint', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('c09d19f5ac172e35dbeb7b279d54de73f7e997e49ca812e446fa362a43b71b58')
       ]);
       expect(tx.body.mint!).toBeDefined();
     });
     it('has StakeKeyRegistration cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('bcd165c882dc17a416bfef7053f0e1cfc3d715f8d7fc05a9803309f795878d9b')
       ]);
       expect(tx.body.certificates![0]).toEqual({
@@ -76,7 +76,7 @@ describe('blockfrostWalletProvider', () => {
       });
     });
     it('has StakeDelegation cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('bcd165c882dc17a416bfef7053f0e1cfc3d715f8d7fc05a9803309f795878d9b')
       ]);
       expect(tx.body.certificates![1]).toEqual({
@@ -88,7 +88,7 @@ describe('blockfrostWalletProvider', () => {
       });
     });
     it('has StakeKeyDeregistration cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('150a445f4d1f2e692791daec9c09f32d6c8c25a3f9ca6c7cf14ff8085375aaa0')
       ]);
       expect(tx.body.certificates![0]).toEqual({
@@ -99,7 +99,7 @@ describe('blockfrostWalletProvider', () => {
       });
     });
     it('PoolRegistration poolParameters are not implemented (null)', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('295d5e0f7ee182426eaeda8c9f1c63502c72cdf4afd6e0ee0f209adf94a614e7')
       ]);
       expect(tx.body.certificates![2]).toEqual({
@@ -109,7 +109,7 @@ describe('blockfrostWalletProvider', () => {
       });
     });
     it('has PoolRetirement cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('545ee7080a01aa085be01dffc073020be04ea3283b945c408c0830e9c4f8253c')
       ]);
       expect(tx.body.certificates![0]).toEqual({
@@ -119,7 +119,7 @@ describe('blockfrostWalletProvider', () => {
       });
     });
     it('has MoveInstantaneousRewards cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([
+      const [tx] = await chainHistoryProvider.transactionsByHashes([
         Cardano.TransactionId('24986c0cd3475419bfb44756c27a0e13a6354a4071153f76a78f9b2d1e596089')
       ]);
       expect(tx.body.certificates![0]).toEqual({
@@ -131,63 +131,67 @@ describe('blockfrostWalletProvider', () => {
     });
     // TODO: blocked by https://github.com/input-output-hk/cardano-db-sync/issues/290
     it.skip('has GenesisKeyDelegation cert', async () => {
-      const [tx] = await walletProvider.transactionsByHashes([Cardano.TransactionId('someValue')]);
+      const [tx] = await chainHistoryProvider.transactionsByHashes([Cardano.TransactionId('someValue')]);
       expect(tx.body.certificates![0]).toBeDefined();
     });
   });
 
   describe('transactionsByAddresses', () => {
     it('Shelley address (addr_test1)', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg')
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg')]
+      });
       expect(txs.length).toBeGreaterThanOrEqual(47);
       expect(txs[0].id).toBe('bedfc2ff545ef1ac3cc4d1a06aa67a6d68a663ffb1092f8764390b8a58ef97b4');
       expect(txs[46].id).toBe('f632b491bb481b4d93fa69e1901ebb623a3af65fde500f1b019eaabd4bb2a980');
     });
     it('extended Shelley address (addr_test1)', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address(
-          'addr_test1qph5x6uahxhxyvtqatzj77sjtjmdjycemt5ncjuj2r4e' +
-            'yflkdap42xncd6cazjarce6jh8mx52fcf8ugststvyklj70qhzhe9h'
-        )
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [
+          Cardano.Address(
+            'addr_test1qph5x6uahxhxyvtqatzj77sjtjmdjycemt5ncjuj2r4e' +
+              'yflkdap42xncd6cazjarce6jh8mx52fcf8ugststvyklj70qhzhe9h'
+          )
+        ]
+      });
       expect(txs.length).toBeGreaterThanOrEqual(4);
       expect(txs[0].id).toBe('a01623f7e3fc679c9f369e06ac0cd942740cade30367b24cedace20a430af1cf');
       expect(txs[3].id).toBe('667f714ee9d9975ca4fa0f5451e006d3dafcdafb7342fe288ebcaf17c100a996');
     });
     it('Icarus Byron address (2c)', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address('2cWKMJemoBai9J7kVvRTukMmdfxtjL9z7c396rTfrrzfAZ6EeQoKLC2y1k34hswwm4SVr')
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [Cardano.Address('2cWKMJemoBai9J7kVvRTukMmdfxtjL9z7c396rTfrrzfAZ6EeQoKLC2y1k34hswwm4SVr')]
+      });
       expect(txs.length).toBeGreaterThanOrEqual(1);
       expect(txs[0].id).toBe('2822d491a890b40cd2a22003b81a97f63c2b8c373b1b0b8dfa1598739fe34c06');
     });
     // Failing because returned transactions are grouped by address
     // TODO: update and reenable test when we decide behaviour
     it.skip('multiple address types', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address('2cWKMJemoBai9J7kVvRTukMmdfxtjL9z7c396rTfrrzfAZ6EeQoKLC2y1k34hswwm4SVr'),
-        Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg'),
-        Cardano.Address(
-          'addr_test1qph5x6uahxhxyvtqatzj77sjtjmdjycemt5ncjuj2r4e' +
-            'yflkdap42xncd6cazjarce6jh8mx52fcf8ugststvyklj70qhzhe9h'
-        )
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [
+          Cardano.Address('2cWKMJemoBai9J7kVvRTukMmdfxtjL9z7c396rTfrrzfAZ6EeQoKLC2y1k34hswwm4SVr'),
+          Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg'),
+          Cardano.Address(
+            'addr_test1qph5x6uahxhxyvtqatzj77sjtjmdjycemt5ncjuj2r4e' +
+              'yflkdap42xncd6cazjarce6jh8mx52fcf8ugststvyklj70qhzhe9h'
+          )
+        ]
+      });
       expect(txs.length).toBeGreaterThanOrEqual(52);
       expect(txs[0].id).toBe('2822d491a890b40cd2a22003b81a97f63c2b8c373b1b0b8dfa1598739fe34c06');
       expect(txs[51].id).toBe('667f714ee9d9975ca4fa0f5451e006d3dafcdafb7342fe288ebcaf17c100a996');
     });
     it('Shelley address not used - no transactions', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address('addr_test1vrfxjeunkc9xu8rpnhgkluptaq0rm8kyxh8m3q9vtcetjwshvpnsm')
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [Cardano.Address('addr_test1vrfxjeunkc9xu8rpnhgkluptaq0rm8kyxh8m3q9vtcetjwshvpnsm')]
+      });
       expect(txs.length).toBe(0);
     });
     it('query ignores invalid transaction (script failure)', async () => {
-      const txs = await walletProvider.transactionsByAddresses([
-        Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg')
-      ]);
+      const txs = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [Cardano.Address('addr_test1vr8nl4u0u6fmtfnawx2rxfz95dy7m46t6dhzdftp2uha87syeufdg')]
+      });
       expect(
         txs.find(
           (tx) => tx.id === Cardano.TransactionId('43149210cbbfbc92bc2b199bb14cb15330414e2288ac31be92a3b5a490f9abfc')
@@ -204,7 +208,10 @@ describe('blockfrostWalletProvider', () => {
       const address = Cardano.Address(
         'addr_test1qp88yvfup4eykezr2dytygwyglfzflyn32dh83ftxkzeg4jrdz3th865e0s2cm6xuzc4xkd8desmtu3p5jfmzkazmxwsm2tk5a'
       );
-      const transactions = await walletProvider.transactionsByAddresses([address], 3_348_548);
+      const transactions = await chainHistoryProvider.transactionsByAddresses({
+        addresses: [address],
+        sinceBlock: 3_348_548
+      });
       expect(transactions[0].id).toBe('264ad5454078db439532e81a5918930779562601b098d6aeae556f785d35e187');
       expect(transactions.length).toBeGreaterThanOrEqual(4);
     });
