+++ packages/wallet/src/services/DelegationTracker/DelegationTracker.ts
@@ -1,9 +1,9 @@
 import {
   Cardano,
+  ChainHistoryProvider,
   SlotEpochCalc,
   StakePoolProvider,
   TimeSettings,
-  WalletProvider,
   createSlotEpochCalc
 } from '@cardano-sdk/core';
 import { DelegationTracker, TransactionsTracker } from '../types';
@@ -26,8 +26,9 @@ import { coldObservableProvider } from '../util';
 import { transactionsWithCertificates } from './transactionCertificates';
 
 export const createBlockEpochProvider =
-  (walletProvider: WalletProvider, retryBackoffConfig: RetryBackoffConfig) => (blockHashes: Cardano.BlockId[]) =>
-    coldObservableProvider(() => walletProvider.blocksByHashes(blockHashes), retryBackoffConfig).pipe(
+  (chainHistoryProvider: ChainHistoryProvider, retryBackoffConfig: RetryBackoffConfig) =>
+  (blockHashes: Cardano.BlockId[]) =>
+    coldObservableProvider(() => chainHistoryProvider.blocksByHashes(blockHashes), retryBackoffConfig).pipe(
       map((blocks) => blocks.map(({ epoch }) => epoch))
     );
