+++ packages/wallet/test/e2e/config.ts
@@ -2,6 +2,7 @@ import * as envalid from 'envalid';
 import {
   BlockFrostAPI,
   blockfrostAssetProvider,
+  blockfrostChainHistoryProvider,
   blockfrostNetworkInfoProvider,
   blockfrostTxSubmitProvider,
   blockfrostUtxoProvider,
@@ -30,10 +31,12 @@ const utxoProviderOptions = ['blockfrost'];
 const walletProviderOptions = ['blockfrost'];
 const assetProviderOptions = ['blockfrost'];
 const keyAgentOptions = ['InMemory', 'Ledger', 'Trezor'];
+const chainHistoryProviderOptions = ['blockfrost'];
 
 const env = envalid.cleanEnv(process.env, {
   ASSET_PROVIDER: envalid.str({ choices: assetProviderOptions }),
   BLOCKFROST_API_KEY: envalid.str(),
+  CHAIN_HISTORY_PROVIDER: envalid.str({ choices: chainHistoryProviderOptions }),
   KEY_AGENT: envalid.str({ choices: keyAgentOptions }),
   LOGGER_MIN_SEVERITY: envalid.str({ choices: loggerMethodNames as string[], default: 'info' }),
   MNEMONIC_WORDS: envalid.makeValidator<string[]>((input) => {
@@ -67,7 +70,8 @@ const blockfrostApi = [
   env.WALLET_PASSWORD,
   env.TX_SUBMIT_PROVIDER,
   env.ASSET_PROVIDER,
-  env.NETWORK_INFO_PROVIDER
+  env.NETWORK_INFO_PROVIDER,
+  env.CHAIN_HISTORY_PROVIDER
 ].includes('blockfrost')
   ? (async () => {
       logger.debug('WalletProvider:blockfrost - Initializing');
@@ -89,7 +93,7 @@ export const assetProvider = (async () => {
   if (env.ASSET_PROVIDER === 'blockfrost') {
     return blockfrostAssetProvider(await blockfrostApi!);
   }
-  throw new Error(`NETWORK_INFO_PROVIDER unsupported: ${env.NETWORK_INFO_PROVIDER}`);
+  throw new Error(`ASSET_PROVIDER unsupported: ${env.ASSET_PROVIDER}`);
 })();
 
 export const txSubmitProvider = (async () => {
@@ -198,6 +202,13 @@ export const utxoProvider = (async () => {
     return blockfrostUtxoProvider(await blockfrostApi!);
   }
   throw new Error(`UTXO_PROVIDER unsupported: ${env.UTXO_PROVIDER}`);
+})();
+
+export const chainHistoryProvider = (async () => {
+  if (env.CHAIN_HISTORY_PROVIDER === 'blockfrost') {
+    return blockfrostChainHistoryProvider(await blockfrostApi!);
+  }
+  throw new Error(`CHAIN_HISTORY_PROVIDER unsupported: ${env.CHAIN_HISTORY_PROVIDER}`);
 })();
 
 export const poolId1 = Cardano.PoolId(env.POOL_ID_1);
