+++ packages/wallet/test/mocks/index.ts
@@ -1,3 +1,4 @@
+export * from './mockData';
 export * from './mockWalletProvider';
 export * from './mockWalletProvider2';
 export * from './mockTxSubmitProvider';
@@ -5,3 +6,4 @@ export * from './testKeyAgent';
 export * from './mockAssetProvider';
 export * from './mockNetworkInfoProvider';
 export * from './mockUtxoProvider';
+export * from './mockChainHistoryProvider';
