+++ packages/web-extension/e2e/.env.example
@@ -5,6 +5,7 @@ ASSET_PROVIDER=blockfrost
 UTXO_PROVIDER=blockfrost
 STAKE_POOL_PROVIDER=stub
 NETWORK_INFO_PROVIDER=blockfrost
+CHAIN_HISTORY_PROVIDER=blockfrost
 BLOCKFROST_API_KEY=testnetNElagmhpQDubE6Ic4XBUVJjV5DROyijO
 NETWORK_ID=0
 MNEMONIC_WORDS="actor scout worth mansion thumb device mass pave gospel secret height document merge text broom kind lesson invest across estate erase interest end century"
